Robust handling of implicit shape operators.
In operator's __init__, if shapein == shapeout, make it square, Do not call reshapein/out if the operator is square. Simplify shapein/out, reshapein/out, _reshapein/out for CompositeOperator. Raise an error if the operator is implicit, non-square and without specialised reshapein/out methods. Raise an error if shapeout is not None and shapein is None in __init__